Refactor views to use more partials
- Use css-grid for grid-layout on account page
- Create account controller
- Remove ActionMailbox
- Use same "empty" views for projects, customers and invoices.
- Generate some fake invoice data.
