+++ app/views/accounts/_projects.html.erb
@@ -1,0 +1,36 @@
+<% projects ||= [] %>
+<div class="col-span-2 lg:col-span-1">
+  <div class="h-full flex flex-col bg-white border-t border-b sm:rounded sm:border shadow">
+    <div class="w-full border-b">
+      <div class="flex justify-between px-6 -mb-px">
+        <h3 class="text-blue-700 py-4 font-normal text-lg">Recent projects</h3>
+        <div class="flex">
+          <button type="button" class="appearance-none py-4 text-blue-700 border-b border-blue-700 mr-3">
+            Active
+          </button>
+          <button type="button" class="appearance-none py-4 text-gray-700 border-b border-transparent hover:border-gray-700">
+            Archived
+          </button>
+        </div>
+      </div>
+    </div>
+    <% if projects.empty? %>
+      <%= render('projects/empty') %>
+    <% else %>
+      <div class="flex flex-1 items-start">
+        <% projects.each do |project| %>
+          <%= render('projects/status_row', project: project) %>
+        <% end %>
+      </div>
+      <div class="border-t px-6 py-3">
+        <div class="text-center text-gray flex justify-center">
+          <div class="flex hover:border-gray-400 hover:text-gray-400 font-medium">
+            <%= link_to(projects_path, class: 'bg-blue-600 border-blue-600 border hover:bg-blue-500 hover:border-blue-500 py-2 px-3 rounded text-white') do %>
+              <%= t('.all_projects') %>
+            <% end %>
+          </div>
+        </div>
+      </div>
+    <% end %>
+  </div>
+</div>
