+++ app/views/accounts/_stats.html.erb
@@ -1,0 +1,91 @@
+<div class="col-span-2 bg-white border-t border-b sm:border-l sm:border-r sm:rounded shadow mb-6">
+  <div class="border-b px-6">
+    <div class="flex justify-between -mb-px">
+      <div class="lg:hidden text-blue-600 py-4 text-lg">
+        Statistics
+      </div>
+      <div class="hidden lg:flex">
+        <button type="button" class="appearance-none py-4 text-blue-600 border-b border-blue-700 mr-6">
+          Projects
+        </button>
+        <button type="button" class="appearance-none py-4 text-gray-700 border-b border-transparent hover:border-gray-700 mr-6">
+          Time logs
+        </button>
+        <button type="button" class="appearance-none py-4 text-gray-700 border-b border-transparent hover:border-gray-700 mr-6">
+          Invoices
+        </button>
+      </div>
+      <div class="flex text-sm">
+        <button type="button" class="appearance-none py-4 text-gray-700 border-b border-transparent hover:border-gray-700 mr-3">
+          1 D
+        </button>
+        <button type="button" class="appearance-none py-4 text-gray-700 border-b border-transparent hover:border-gray-700 mr-3">
+          1 W
+        </button>
+        <button type="button" class="appearance-none py-4 text-blue-700 border-b border-blue-700 mr-3">
+          1 M
+        </button>
+        <button type="button" class="appearance-none py-4 text-gray-700 border-b border-transparent hover:border-gray-700 mr-3">
+          1 Y
+        </button>
+        <button type="button" class="appearance-none py-4 text-gray-700 border-b border-transparent hover:border-gray-700">
+          Total
+        </button>
+      </div>
+    </div>
+  </div>
+  <div class="flex items-center px-6 lg:hidden">
+    <div class="w-3/4 py-6">
+      <div class="text-gray-900 mb-2">
+        <span class="text-5xl">123</span>
+      </div>
+      <div class="text-green-500 text-sm">
+        &uarr; 5 more than last month
+      </div>
+    </div>
+    <div class="flex-grow py-6">
+      <div class="relative">
+        <select class="block appearance-none w-full bg-white border border-gray-200 px-4 py-2 pr-8 rounded" id="grid-state">
+          <option>Projects</option>
+          <option>Time logs</option>
+          <option>Invoices</option>
+        </select>
+        <div class="pointer-events-none absolute inset-y-0 right-0 flex items-center px-2 text-gray-700">
+          <svg class="fill-current h-4 w-4" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20"><path d="M9.293 12.95l.707.707L15.657 8l-1.414-1.414L10 10.828 5.757 6.586 4.343 8z"/></svg>
+        </div>
+      </div>
+    </div>
+  </div>
+  <div class="hidden lg:flex">
+    <div class="w-1/3 text-center py-8">
+      <div class="border-r">
+        <div class="text-gray-900 mb-2">
+          <span class="text-5xl"><%= @counts[:active] %></span>
+        </div>
+        <div class="text-sm uppercase text-gray tracking-wide">
+          Active
+        </div>
+      </div>
+    </div>
+    <div class="w-1/3 text-center py-8">
+      <div class="border-r">
+        <div class="text-gray-900 mb-2">
+          <span class="text-5xl"><%= @counts[:archived] %></span>
+        </div>
+        <div class="text-sm uppercase text-gray tracking-wide">
+          Archived
+        </div>
+      </div>
+    </div>
+    <div class="w-1/3 text-center py-8">
+      <div>
+        <div class="text-gray-900 mb-2">
+          <span class="text-5xl"><%= @counts[:total] %></span>
+        </div>
+        <div class="text-sm uppercase text-gray tracking-wide">
+          Total
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
