+++ app/views/accounts/_tasks.html.erb
@@ -1,0 +1,10 @@
+<div class="col-span-2 lg:col-span-1">
+  <div class="bg-white border-t border-b sm:rounded sm:border shadow">
+    <div class="border-b">
+      <div class="flex justify-between px-6 -mb-px">
+        <h3 class="text-blue-700 py-4 font-normal text-lg">Recent time logs</h3>
+      </div>
+    </div>
+    <%= render('tasks/empty') %>
+  </div>
+</div>
