+++ app/views/accounts/settings.html.erb
@@ -1,3 +1,7 @@
+<% content_for(:sidebar) do %>
+  <%= render('application/account_menu', current: :settings) %>
+<% end %>
+
 <div class="w-full overflow-y-auto lg:flex-grow px-4">
   <div class="flex-grow sm:px-4 pt-2 md:pt-6 max-w-6xl">
     <div class="w-full mb-12">
@@ -132,7 +136,7 @@
           <div class="flex flex-wrap -mx-3">
             <div class="w-full px-3">
               <label class="tracking-wide text-gray-900 text-base font-semibold">
-                About 
+                About
                 <p class="font-normal text-gray-700 inline">(Describe your company)</p>
               </label>
               <textarea class="appearance-none mt-2 block w-full bg-gray-100 text-gray-700 border border-gray-400 rounded py-3 px-4 mb-3 leading-tight focus:outline-none focus:bg-white focus:border-gray-500 h-32" id="subject" type="description"></textarea>
@@ -162,7 +166,7 @@
           <div class="flex flex-wrap -mx-3">
             <div class="w-full flex-wrap px-3 mb-3">
               <label class="tracking-wide text-gray-800 text-base font-semibold">
-                Status 
+                Status
                 <p class="font-normal text-gray-700 inline">(Use this if you are unable to accept more inquries at the moment)</p>
               </label>
               <span class="border rounded-full border-grey-400 flex items-center cursor-pointer w-12 bg-green-500 justify-end mt-2">
