+++ app/views/accounts/show.html.erb
@@ -1,0 +1,13 @@
+<% content_for(:sidebar) do %>
+  <%= render('application/account_menu', current: :account) %>
+<% end %>
+
+<div class="grid grid-cols-2 gap-4 w-full px-8 pt-2 lg:pt-6 overflow-y-auto">
+  <div class="col-span-2 mb-8">
+    <h1 class="text-3xl font-bold text-gray-800"><%= t(".title") %></h1>
+  </div>
+  <%= render("stats") %>
+  <%= render("projects", projects: @projects) %>
+  <%= render("tasks") %>
+  <%= render("invoices") %>
+</div>
