+++ app/views/customers/_empty.html.erb
@@ -1,13 +1,13 @@
-<div class="w-4/5 mx-auto text-center p-6 py-16">
-  <div class="w-2/3 lg:w-1/3 mx-auto mt-4 mb-4">
+<div class="text-center p-6">
+  <div class="w-24 mx-auto mt-4 mb-4">
     <%= image_tag('add_first_customer.svg', alt: 'Builders')%>
   </div>
-  <p class="text-2xl text-gray-900 font-medium mb-4">No customers found!</p>
-  <p class="text-gray max-w-xs mx-auto mb-6">To register your first customer, click the button below.</p>
+  <p class="text-2xl text-gray-900 font-medium mb-4"><%= t(".no_customers_found") %></p>
+  <p class="text-gray max-w-xs mx-auto mb-6"><%= t(".register_first_customer") %></p>
   <div class="text-center text-gray flex justify-center mt-8">
     <div class="flex hover:border-gray-400 hover:text-gray-400 font-medium">
       <%= link_to(new_customer_path, class: 'bg-blue-600 border-blue-600 border hover:bg-blue-500 hover:border-blue-500 py-2 px-3 rounded text-white') do %>
-        <%= t('.new_customer!') %>
+        <%= t('.new_customer') %>
       <% end %>
     </div>
   </div>
