+++ app/views/customers/_row.html.erb
@@ -1,30 +1,33 @@
-<%= link_to(customer_path(customer), class: 'w-full text-left hover:bg-blue-200') do %>
-  <div class="w-full flex py-4">
-    <div class="w-20 md:w-24 flex justify-center items-center">
-      <div
-      class="rounded-full h-9 w-9 md:h-10 md:w-10 bg-blue-500 inline-flex items-center justify-center text-white font-bold">
-        <%= customer.initials %>
+<% odd ||= false %>
+<div class="flex border-b <%= odd ? "bg-white" : "bg-gray-200" %>">
+  <%= link_to(customer_path(customer), class: 'w-full text-left hover:bg-blue-200') do %>
+    <div class="w-full flex py-4">
+      <div class="w-20 md:w-24 flex justify-center items-center">
+        <div
+        class="rounded-full h-9 w-9 md:h-10 md:w-10 bg-blue-500 inline-flex items-center justify-center text-white font-bold">
+          <%= customer.initials %>
+        </div>
+      </div>
+      <div class="w-full flex items-center">
+        <div class="w-1/3 lg:w-1/4">
+          <p class="text-sm md:text-base"><%= customer.name%></p>
+          <p class="text-sm text-gray-700 opacity-75">Individual</p>
+        </div>
+        <div class="w-1/3 lg:w-1/4">
+          <p class="text-sm md:text-base"><%= customer.latest_project %></p>
+        </div>
+        <div class="w-1/4 hidden lg:inline-block">
+          <p class="text-sm md:text-base"><%= customer.projects.count %></p>
+        </div>
+        <div class="w-1/3 lg:w-1/4 flex justify-center md:justify-start">
+          5.799,00 kr
+        </div>
+      </div>
+      <div class="w-16 hidden md:flex items-center">
+        <span class="icon text-gray-600">
+          <i class="fas fa-lg fa-edit"></i>
+        </span>
       </div>
     </div>
-    <div class="w-full flex items-center">
-      <div class="w-1/3 lg:w-1/4">
-        <p class="text-sm md:text-base"><%= customer.name%></p>
-        <p class="text-sm text-gray-700 opacity-75">Individual</p>
-      </div>
-      <div class="w-1/3 lg:w-1/4">
-        <p class="text-sm md:text-base"><%= customer.latest_project %></p>
-      </div>
-      <div class="w-1/4 hidden lg:inline-block">
-        <p class="text-sm md:text-base"><%= customer.projects.count %></p>
-      </div>
-      <div class="w-1/3 lg:w-1/4 flex justify-center md:justify-start">
-        5.799,00 kr
-      </div>
-    </div>
-    <div class="w-16 hidden md:flex items-center">
-      <span class="icon text-gray-600">
-        <i class="fas fa-lg fa-edit"></i>
-      </span>
-    </div>
-  </div>
-<% end %>
+  <% end %>
+</div>
