+++ app/views/invoices/_status_row.html.erb
@@ -1,4 +1,5 @@
 <% odd ||= false %>
+<% date ||= Time.zone.now - rand(-100..100).days %>
 <div class="flex-grow flex px-6 py-2 text-gray-900 border-b -mx-4 <%= odd ? "bg-gray-300" : "bg-white" %>">
   <div class="w-1/4 px-4 flex flex-wrap items-center justify-start">
     <div class="">
@@ -13,6 +14,6 @@
     <p class="text-md"><%= rand(5000..12000) %> SEK</p>
   </div>
   <div class="w-1/4 flex items-center justify-center">
-    <p class="text-md text-green-500"><%= l(Time.zone.now() - rand(-100..100).days, format: "%Y-%m-%d") %></p>
+    <p class="text-md <%= date < Time.zone.now ? "text-red-500" : "text-green-500"%>"><%= l(date, format: "%Y-%m-%d") %></p>
   </div>
 </div>
