+++ app/views/projects/_row.html.erb
@@ -1,38 +1,41 @@
-<%= link_to(project_path(project), class: 'w-full text-left hover:bg-blue-200') do %>
-  <div class="w-full flex py-4">
-    <div class="w-20 md:w-24 flex justify-center items-center">
-      <div class="rounded-full h-9 w-9 md:h-10 md:w-10 bg-blue-500 inline-flex
-                  items-center justify-center text-white font-bold">
-        <i class="fas fa-lg <%= project.symbol %>"></i>
-      </div>
-    </div>
-    <div class="w-full flex items-center">
-      <div class="w-1/3 lg:w-1/4">
-        <p class="text-sm md:text-base"><%= project.title %></p>
-        <p class="text-sm text-gray-700 opacity-75"><%= project.kind %></p>
-      </div>
-      <div class="w-1/3 lg:w-1/4">
-        <p class="text-sm md:text-base"><%= project.customer.name %></p>
-        <p class="text-sm text-gray-700 opacity-75">Privatperson</p>
-      </div>
-      <div class="w-1/4 hidden lg:inline-block">
-        <p class="text-sm md:text-base">Genvägen 75a</p>
-        <p class="text-sm text-gray-700 opacity-75">Helsingborg</p>
-      </div>
-      <div class="w-1/3 lg:w-1/4 flex justify-center md:justify-start">
-        <div class="w-1/3 lg:w-1/4 flex justify-center md:justify-start">
-          <span class="text-sm md:text-base inline-flex bg-gray-200 text-gray-800
-                       font-semibold rounded-full h-6 px-3 md:px-4 justify-center
-                       items-center">
-            Archived
-          </span>
+<% odd ||= false %>
+<div class="flex border-b <%= odd ? "bg-white" : "bg-gray-200" %>">
+  <%= link_to(project_path(project), class: 'w-full text-left hover:bg-blue-200') do %>
+    <div class="w-full flex py-4">
+      <div class="w-20 md:w-24 flex justify-center items-center">
+        <div class="rounded-full h-9 w-9 md:h-10 md:w-10 bg-blue-500 inline-flex
+                    items-center justify-center text-white font-bold">
+          <i class="fas fa-lg <%= project.symbol %>"></i>
         </div>
       </div>
+      <div class="w-full flex items-center">
+        <div class="w-1/3 lg:w-1/4">
+          <p class="text-sm md:text-base"><%= project.title %></p>
+          <p class="text-sm text-gray-700 opacity-75"><%= project.kind %></p>
+        </div>
+        <div class="w-1/3 lg:w-1/4">
+          <p class="text-sm md:text-base"><%= project.customer.name %></p>
+          <p class="text-sm text-gray-700 opacity-75">Privatperson</p>
+        </div>
+        <div class="w-1/4 hidden lg:inline-block">
+          <p class="text-sm md:text-base">Genvägen 75a</p>
+          <p class="text-sm text-gray-700 opacity-75">Helsingborg</p>
+        </div>
+        <div class="w-1/3 lg:w-1/4 flex justify-center md:justify-start">
+          <div class="w-1/3 lg:w-1/4 flex justify-center md:justify-start">
+            <span class="text-sm md:text-base inline-flex bg-gray-200 text-gray-800
+                        font-semibold rounded-full h-6 px-3 md:px-4 justify-center
+                        items-center">
+              Archived
+            </span>
+          </div>
+        </div>
+      </div>
+      <div class="w-16 hidden md:flex items-center">
+        <span class="icon text-gray-600">
+          <i class="fas fa-lg fa-edit"></i>
+        </span>
+      </div>
     </div>
-    <div class="w-16 hidden md:flex items-center">
-      <span class="icon text-gray-600">
-        <i class="fas fa-lg fa-edit"></i>
-      </span>
-    </div>
-  </div>
-<% end %>
+  <% end %>
+</div>
