+++ app/views/projects/_status_row.html.erb
@@ -1,4 +1,4 @@
-<%= link_to(project_path(project), class: "flex-grow flex px-6 py-4 text-gray-900 items-center border-b -mx-4") do %>
+<%= link_to(project_path(project), class: "flex w-full px-6 py-4 text-gray-900 items-center border-b") do %>
   <div class="w-2/5 px-4 flex items-center">
     <div class="rounded-full h-10 w-10 bg-indigo-500 inline-flex mr-3 items-center justify-center">
       <span class="icon text-white">
