+++ app/views/projects/index.html.erb
@@ -1,15 +1,14 @@
 <% content_for(:sidebar) do %>
   <%= render('application/account_menu', current: :projects) %>
 <% end %>
-
 <div class="w-full overflow-y-auto lg:flex-grow px-4">
   <div class="flex-grow mx-auto sm:px-4 pt-2 md:pt-6">
     <div class="w-full flex justify-between items-center">
       <h1 class="text-3xl font-bold text-gray-800">Projects</h1>
       <div class="flex items-end md:items-center mt-2">
         <%= link_to(new_project_path, class: 'bg-blue-200 border border-blue-200 hover:bg-blue-500 hover:border-blue-500 py-2 px-3 rounded text-blue-800 font-semibold') do %>
-        <%= icon('fas', 'plus', class: 'mr-1 text-blue-900 text-sm') %>
-        <%= t('.new_project') %>
+          <%= icon('fas', 'plus', class: 'mr-1 text-blue-900 text-sm') %>
+          <%= t('.new_project') %>
         <% end %>
       </div>
     </div>
@@ -104,33 +103,12 @@
           </div>
         </div>
       </div>
-      <% if (Project.all.count < 1) %>
-      <div class="w-4/5 mx-auto text-center p-6 py-16">
-        <div class="w-2/3 lg:w-1/3 mx-auto mt-4 mb-4">
-          <%= image_tag('add_first_customer.svg', alt: 'Builders')%>
-        </div>
-        <p class="text-2xl text-gray-900 font-medium mb-4">No projects found!</p>
-        <p class="text-gray max-w-xs mx-auto mb-6">To register your first project, click the button below.</p>
-        <div class="text-center text-gray flex justify-center mt-8">
-          <div class="flex hover:border-gray-400 hover:text-gray-400 font-medium">
-            <%= link_to(new_project_path, class: 'bg-blue-600 border-blue-600 border hover:bg-blue-500 hover:border-blue-500 py-2 px-3 rounded text-white') do %>
-            <%= t('.new_project!') %>
-            <% end %>
-          </div>
-        </div>
-      </div>
+      <% if @projects.empty? %>
+        <%= render('projects/empty') %>
       <% else %>
-      <% @projects.each_with_index do |project, idx| %>
-      <% if (idx % 2 == 0) %>
-      <div class="flex border-b bg-white">
-        <%= render("row", project: project) %>
-      </div>
-      <% else %>
-      <div class="flex border-b bg-gray-200">
-        <%= render("row", project: project) %>
-      </div>
-      <% end %>
-      <% end %>
+        <% @projects.each_with_index do |project, idx| %>
+          <%= render("row", project: project, odd: idx % 2 == 0) %>
+        <% end %>
       <% end %>
       <div class="px-3 bg-white border-t flex flex-col xs:flex-row items-end justify-center">
         <div class="px-2 align-middle inline-flex flex items-center">
